+++ janusgraph-es/src/main/java/org/janusgraph/diskstorage/es/ElasticMajorVersion.java
@@ -23,6 +23,8 @@ public enum ElasticMajorVersion {
 
     SEVEN(7),
 
+    EIGHT(8),
+
     ;
 
     static final Pattern PATTERN = Pattern.compile("(\\d+)\\.\\d+\\.\\d+.*");
@@ -44,6 +46,8 @@ public enum ElasticMajorVersion {
                 return ElasticMajorVersion.SIX;
             case 7:
                 return ElasticMajorVersion.SEVEN;
+            case 8:
+                return ElasticMajorVersion.EIGHT;
             default:
                 throw new IllegalArgumentException("Unsupported Elasticsearch server major version: " + value);
         }